@@ -6,7 +6,7 @@ description = ""
 authors = ["Xinyuan Yao <yao.ntno@gmail.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 torch = "^1.7.0"
 scipy = "^1.5.2"
 tqdm = "^4.49.0"
@@ -18,11 +18,11 @@ opencv-python = "^4.4.0"
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
 black = "^20.8b1"
-mypy = "^0.790"
 pytest-cov = "^2.10.1"
 pytest-benchmark = "^3.2.3"
 torch-optimizer = "^0.1.0"
 torchnet = "^0.0.4"
+mypy = "^0.800"
 
 [build-system]
 requires = ["poetry>=0.12"]
