@@ -1,6 +1,5 @@
-from typing import Tuple, List, Callable
 from .box import *
 from .image import *
 from typing_extensions import Literal
 
-FP = Tuple[Tensor, Tensor, Tensor, Tensor, Tensor]  # p3, p4, p5, p6, p7
+FP = tuple[Tensor, Tensor, Tensor, Tensor, Tensor]  # p3, p4, p5, p6, p7
