@@ -1,10 +1,9 @@
-import typing as t
-from typing import NewType, Tuple
+from typing import NewType
 from torch import Tensor, ByteTensor, FloatTensor
 
 ImageId = NewType("ImageId", str)
 Image = NewType("Image", Tensor)  # [C, H, W] dtype
 
 ImageBatch = NewType("ImageBatch", Tensor)  # [B, C, H, W]
-ImageSize = Tuple[int, int]  # W, H
-RGB = Tuple[int, int, int]
+ImageSize = tuple[int, int]  # W, H
+RGB = tuple[int, int, int]
