@@ -1,5 +1,4 @@
 import numpy as np, torch
-from typing import Tuple, List, Dict
 from object_detection.metrics.average_precision import AveragePrecision
 from object_detection.entities import Boxes, Labels, Confidences
 
@@ -36,6 +35,6 @@ class MeanAveragePrecision:
             )
 
     @torch.no_grad()
-    def __call__(self) -> Tuple[float, Dict[int, float]]:
+    def __call__(self) -> tuple[float, dict[int, float]]:
         aps = {k: v() for k, v in self.aps.items()}
         return np.fromiter(aps.values(), dtype=float).mean(), aps
